Allow decimal grades and add a two-exam results sheet
Grades (Grade.Value) are now decimal instead of integer, so teachers can
enter fractional scores like 85.5. A custom model binder accepts both
"85.5" and "85,5" (comma as decimal separator) regardless of the
request's display language, since teachers naturally type grades that
way and HTML number inputs handle locale inconsistently.

Also fixes a leftover bug from the earlier 1-5 -> 0-100 grade scale
migration: the group journal's grade input still validated against the
old Range(1, 5), silently rejecting any 100-point grade on save.

Added an exam sheet feature matching the two-exam register teachers
already keep by hand: for a group + subject, each student gets Exam 1,
Exam 2 and an auto-computed average. Teachers manage their own
group/subject sheets under Teacher/Exams; Admin has the same view for
any group/subject under Admin/Exams.

+++ Areas/Admin/Controllers/DashboardController.cs
@@ -128,7 +128,7 @@ public class DashboardController : Controller
             GroupsCount = await _context.Groups.CountAsync(),
             SubjectsCount = await _context.Subjects.CountAsync(),
             TodayLessonsCount = todayLessons.Count,
-            AverageGrade = allGrades.Count == 0 ? 0 : allGrades.Average(),
+            AverageGrade = allGrades.Count == 0 ? 0 : (double)allGrades.Average(),
             AttendanceRate = allAttendance.Count == 0
                 ? 0
                 : allAttendance.Count(a => a != AttendanceStatus.Absent) * 100.0 / allAttendance.Count,
